pyproject.toml
[project]
name = "doc-layout-forge"
version = "0.1.0"
description = "FastAPI service for document layout detection and evaluation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "doclayout-yolo>=0.0.4",
    "fastapi>=0.117.1",
    "huggingface-hub>=0.35.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.36.0",
]

[dependency-groups]
dev = [
    "aiofiles>=24.1.0",
    "ipykernel>=6.30.1",
    "pre-commit>=4.3.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
plugins = ["pydantic.mypy"]
mypy_path = ["src"]
namespace_packages = true
explicit_package_bases = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unreachable = true
show_error_codes = true
pretty = true
strict = true
disallow_untyped_decorators = false
strict_optional = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.ruff]
# match black
line-length = 100
target-version = "py313"
lint.select = [
    "ALL", # include all the rules, including new ones
]
lint.ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D203",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "G004",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "EM102",
    "N803",
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "T201",
    "FBT001",
    "FBT002",
    "FBT003",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PGH003",
    "EM101",
    "RUF001",
]
